Redeem cleanup system
Added an attempt to prevent "orphaned redeems," or redeems that were added by a previous run of the application that the current version doesn't recognize or need. The list of TWINGE-managed redeems is already compared to the list of redeem nodes in the tree to assign them their IDs, so now we also remove those from a dictionary as we do so, and then remove any redeems that were not "claimed" in the process under the assumption those redeems do not exist any more.